import generateRandomPosition function into eggJar and enemy.js and update spawn code, remove function and unnessary imports from main.js, move utils folder into js folder
