+++ assets/js/eggJar.js
@@ -1,8 +1,11 @@
+import generateRandomPosition from "./utils/generateRandomPosition.js";
+
 /**
  * creates a static jar of eggs - the function is
  * imported and called in the main.js file
  */
-const spawnEggJar = (randPosX, randPosY, colour) => {
+const spawnEggJar = (colour) => {
+  const [randPosX, randPosY] = generateRandomPosition(); // using array destructuring assign variables to the return values
   const eggJar = add([
     rect(40, 40), // placeholder until we have a sprite
     pos(randPosX, randPosY),
