+++ assets/js/main.js
@@ -1,6 +1,4 @@
-import spawnEnemy from "./enemy.js";
 import generateScenes from "./scenes.js";
-import spawnEggJar from "./eggJar.js";
 
 // initialize kaboom context
 kaboom({
@@ -28,26 +26,6 @@ const player = add([
   color(255, 0, 0),
   "player",
 ]);
-
-/**
- * function to generate a random position that is
- * a set distance away from the player
- */
-const generateRandomPosition = (playerInitialXPos, playerInitialYPos) => {
-  const randomXPos = Math.floor(Math.random() * width());
-  const randomYPos = Math.floor(Math.random() * height());
-  const distanceFromPlayer = 200;
-  if (
-    randomXPos > playerInitialXPos - distanceFromPlayer &&
-    randomXPos < playerInitialXPos + distanceFromPlayer &&
-    randomYPos > playerInitialYPos - distanceFromPlayer &&
-    randomYPos < playerInitialYPos + distanceFromPlayer
-  ) {
-    return generateRandomPosition(playerInitialXPos, playerInitialYPos);
-  } else {
-    return [randomXPos, randomYPos];
-  }
-};
 
 // set the initial player speed
 let playerSpeed = 120;
